weatherbot-dev/requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
tornado==6.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
zipp==3.6.0
